[pickers] Add support for single input fields in range pickers
